build: structure protobuf for kmp project
