+++ buildSrc/src/main/kotlin/io/embrace/otel/BuildPlugin.kt
@@ -2,6 +2,7 @@ package io.embrace.opentelemetry.kotlin
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 
 class BuildPlugin : Plugin<Project> {
@@ -13,6 +14,11 @@ class BuildPlugin : Plugin<Project> {
             project.configureDetekt()
             project.configureBinaryCompatValidation()
             project.configureExplicitApiMode(kotlin)
+        }
+        project.pluginManager.withPlugin("org.jetbrains.kotlin.jvm") {
+            val kotlin = project.extensions.getByType(KotlinJvmProjectExtension::class.java)
+            project.configureDetekt()
+            project.configureBinaryCompatValidation()
         }
         project.configurePublishing()
     }
