+++ config/detekt/detekt.yml
@@ -83,3 +83,7 @@ custom-rules:
     allowedFileIds:
       - "io.embrace.opentelemetry.kotlin.aliases.Aliases"
       - "io.embrace.opentelemetry.kotlin.factory.SpanFactoryImpl"
+      - "io.embrace.opentelemetry.kotlin.export.ProtobufSerialization.jvm"
+      - "io.embrace.opentelemetry.kotlin.export.OtlpClientTest"
+      - "io.embrace.opentelemetry.kotlin.logging.export.OtlpHttpLogRecordExporterTest"
+      - "io.embrace.opentelemetry.kotlin.tracing.export.OtlpHttpSpanExporterTest"
