+++ opentelemetry-kotlin-exporters-jvm/build.gradle.kts
@@ -1,5 +1,7 @@
 import com.google.protobuf.gradle.id
 import de.undercouch.gradle.tasks.download.Download
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+import org.jetbrains.kotlin.gradle.dsl.KotlinVersion
 
 plugins {
     kotlin("jvm")
@@ -57,10 +59,25 @@ protobuf {
     }
 }
 
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
+}
+
 kotlin {
     sourceSets {
         main {
             kotlin.srcDir("build/generated/source/proto/main/kotlin")
         }
     }
+    compilerOptions {
+        jvmTarget.set(JvmTarget.JVM_1_8)
+        apiVersion.set(KotlinVersion.KOTLIN_1_8)
+        languageVersion.set(KotlinVersion.KOTLIN_1_8)
+    }
+}
+
+// Disable Detekt tasks for generated code
+tasks.withType<io.gitlab.arturbosch.detekt.Detekt>().configureEach {
+    enabled = false
 }
