+++ opentelemetry-kotlin-exporters-jvm/gradle.properties
@@ -1,0 +1,3 @@
+io.embrace.opentelemetry.kotlin.containsPublicApi=false
+io.embrace.opentelemetry.kotlin.javaSdkCompatModule=true
+io.embrace.opentelemetry.kotlin.enableCodeCoverage=false
