+++ opentelemetry-kotlin-exporters-jvm/src/test/kotlin/io/embrace/opentelemetry/kotlin/logging/export/ExportLogsServiceRequestTest.kt
@@ -2,9 +2,9 @@ package io.embrace.opentelemetry.kotlin.logging.export
 
 import io.embrace.opentelemetry.kotlin.ExperimentalApi
 import io.embrace.opentelemetry.kotlin.logging.model.FakeReadableLogRecord
+import org.junit.Test
 import kotlin.collections.get
 import kotlin.test.assertEquals
-import org.junit.Test
 
 @OptIn(ExperimentalApi::class)
 class ExportLogsServiceRequestTest {
