+++ opentelemetry-kotlin-exporters/README.md
@@ -1,3 +1,5 @@
 # opentelemetry-kotlin-exporters
 
-This module contains span/log exporter implementations.
+This module contains span/log exporter implementations. Currently only the JVM platform is
+supported. `ProtobufSerialization` contains the functions that would be necessary to
+serialize/deserialize protobuf classes on non-JVM platforms in future.
