+++ opentelemetry-kotlin-exporters/api/opentelemetry-kotlin-exporters.api
@@ -1,0 +1,8 @@
+public final class io/embrace/opentelemetry/kotlin/logging/export/OtlpHttpLogRecordExporterApi_jvmKt {
+	public static final fun createOtlpHttpLogRecordExporter (Ljava/lang/String;)Lio/embrace/opentelemetry/kotlin/logging/export/LogRecordExporter;
+}
+
+public final class io/embrace/opentelemetry/kotlin/tracing/export/OtlpHttpSpanExporterApi_jvmKt {
+	public static final fun createOtlpHttpSpanExporter (Ljava/lang/String;)Lio/embrace/opentelemetry/kotlin/tracing/export/SpanExporter;
+}
+
