+++ opentelemetry-kotlin-exporters/src/appleMain/kotlin/io/embrace/opentelemetry/kotlin/export/HttpClientInstance.apple.kt
@@ -1,0 +1,6 @@
+package io.embrace.opentelemetry.kotlin.export
+
+import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.engine.cio.CIO
+
+internal actual fun createHttpEngine(): HttpClientEngine = CIO.create()
