+++ opentelemetry-kotlin-exporters/src/appleMain/kotlin/io/embrace/opentelemetry/kotlin/logging/export/OtlpHttpLogRecordExporterApi.apple.kt
@@ -1,0 +1,8 @@
+package io.embrace.opentelemetry.kotlin.logging.export
+
+import io.embrace.opentelemetry.kotlin.ExperimentalApi
+
+@OptIn(ExperimentalApi::class)
+public actual fun createOtlpHttpLogRecordExporter(baseUrl: String): LogRecordExporter {
+    throw UnsupportedOperationException()
+}
