+++ opentelemetry-kotlin-exporters/src/appleMain/kotlin/io/embrace/opentelemetry/kotlin/tracing/export/OtlpHttpSpanExporterApi.apple.kt
@@ -1,0 +1,8 @@
+package io.embrace.opentelemetry.kotlin.tracing.export
+
+import io.embrace.opentelemetry.kotlin.ExperimentalApi
+
+@OptIn(ExperimentalApi::class)
+public actual fun createOtlpHttpSpanExporter(baseUrl: String): SpanExporter {
+    throw UnsupportedOperationException()
+}
