+++ opentelemetry-kotlin-exporters/src/commonMain/kotlin/io/embrace/opentelemetry/kotlin/export/OtlpEndpoint.kt
@@ -1,0 +1,6 @@
+package io.embrace.opentelemetry.kotlin.export
+
+internal enum class OtlpEndpoint(val path: String) {
+    Logs("v1/logs"),
+    Traces("v1/traces")
+}
