+++ opentelemetry-kotlin-exporters/src/commonMain/kotlin/io/embrace/opentelemetry/kotlin/logging/export/OtlpHttpLogRecordExporterApi.kt
@@ -1,0 +1,9 @@
+package io.embrace.opentelemetry.kotlin.logging.export
+
+import io.embrace.opentelemetry.kotlin.ExperimentalApi
+
+/**
+ * Creates a log record exporter that sends telemetry to the specified URL over OTLP.
+ */
+@ExperimentalApi
+public expect fun createOtlpHttpLogRecordExporter(baseUrl: String): LogRecordExporter
