+++ opentelemetry-kotlin-exporters/src/commonMain/kotlin/io/embrace/opentelemetry/kotlin/tracing/export/OtlpHttpSpanExporterApi.kt
@@ -1,0 +1,9 @@
+package io.embrace.opentelemetry.kotlin.tracing.export
+
+import io.embrace.opentelemetry.kotlin.ExperimentalApi
+
+/**
+ * Creates a span exporter that sends telemetry to the specified URL over OTLP.
+ */
+@ExperimentalApi
+public expect fun createOtlpHttpSpanExporter(baseUrl: String): SpanExporter
