+++ opentelemetry-kotlin-exporters/src/jvmMain/kotlin/io/embrace/opentelemetry/kotlin/export/HttpClientInstance.jvm.kt
@@ -1,0 +1,6 @@
+package io.embrace.opentelemetry.kotlin.export
+
+import io.ktor.client.engine.HttpClientEngine
+import io.ktor.client.engine.okhttp.OkHttp
+
+internal actual fun createHttpEngine(): HttpClientEngine = OkHttp.create()
