--- opentelemetry-kotlin-exporters/src/main/kotlin/io/embrace/opentelemetry/kotlin/export/ExportDefaults.kt
@@ -1,5 +1,0 @@
-package io.embrace.opentelemetry.kotlin.export
-
-internal const val exportInitialDelayMs: Long = 30_000L // 30s
-internal const val exportMaxAttemptIntervalMs: Long = 600000L // 10 mins
-internal const val exportMaxAttempts: Int = 8 // maximum of 8 retries per export call
