--- opentelemetry-kotlin-exporters/src/main/kotlin/io/embrace/opentelemetry/kotlin/export/OtlpEndpoint.kt
@@ -1,6 +1,0 @@
-package io.embrace.opentelemetry.kotlin.export
-
-enum class OtlpEndpoint(val path: String) {
-    Logs("v1/logs"),
-    Traces("v1/traces");
-}
